move `WalletState` to UI
